// src/renderer/typings.d.ts
export interface IUser {
  id?: string;
  name: string;
  password: string;
  email?: string;
}